descoperire config-ponderi: robustete (din verificare adversariala)
Inchide 3 cai prin care editorul D1 putea brick-ui permanent descoperirea:
- CRITIC: pondere fractionara/non-finita (2.5/NaN/Inf) trecea de validare -> persista -> scor_candidat
  crapa in pydantic -> inghitit ca 502 inselator. Fix: POST respinge non-finit/negativ/non-intreg (422,
  coercit la int); AttributeBreakdown.pondere: float (tolerant, fara crash); fuzioneaza_override ignora non-finit.
- MAJOR: editari partiale cumulate ocoleau garda suma>0 (validata pe default). Fix: validez suma EFECTIVA
  cumulata (override existent + editare) inainte de persistare.
- MAJOR: categorie non-dict pe disc -> salveaza_override crapa cu 500. Fix: reinitializez categoria corupta.
- MINOR: scriere atomica (temp + os.replace, allow_nan=False) + sanitizare la persistare.
+4 teste robustete. 562 verzi, ruff curat.

+++ evaluare-anevar/src/evaluare/discovery/profiles.py
@@ -40,7 +40,7 @@ class AttributeBreakdown(BaseModel):
     valoare_subiect: str | None = None
     valoare_candidat: str | None = None      # textul brut gasit in anunt
     d: float | None = None                   # dissimilaritate [0,1]; None daca necunoscut
-    pondere: int = 0
+    pondere: float = 0                        # tolereaza float (ponderi editate); afisarea ramane intreaga
     contributie: float | None = None         # pondere * d; None daca necunoscut
     cunoscut: bool = True
